lambda_deployment/requirements.txt
"""
Lambda requirements - minimal for serverless
"""
moviepy==1.0.3
requests==2.32.3
boto3==1.26.137
pillow==10.4.0
